+++ CHANGELOG.md
@@ -19,7 +19,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - Forward cmem-plugin-template to v7.1.0
 - Removed accepting parameter settings on input port
-
+- Changed name of parameter "skip_lines" to "skip_rows"
+- Changed Parameter descriptions
 
 
 ## [2.0.0]